+++ apps/server/src/plugins/auth/permissions.ts
@@ -11,6 +11,8 @@ const statement = {
   submissions: ["read", "read_all", "create", "update", "delete"],
   settings: ["read", "update"],
   deleted_entries: ["read"],
+  uke_permits_orphaned: ["read"],
+  uke_permits: ["read"],
 } as const;
 
 export const accessControl = createAccessControl(statement);
@@ -24,6 +26,20 @@ export const userRole = accessControl.newRole({
   submissions: ["read", "create"],
   settings: ["read"],
   deleted_entries: ["read"],
+  uke_permits: ["read"],
+});
+
+export const editorRole = accessControl.newRole({
+  stations: ["create", "delete", "read", "update"],
+  cells: ["create", "delete", "read", "update"],
+  operators: ["read"],
+  bands: ["read"],
+  locations: ["create", "read", "update"],
+  submissions: ["create", "read_all", "read", "update"],
+  settings: ["read"],
+  deleted_entries: ["read"],
+  uke_permits: ["read"],
+  uke_permits_orphaned: ["read"],
 });
 
 export const modRole = accessControl.newRole({
@@ -39,4 +55,7 @@ export const adminRole = accessControl.newRole({
   bands: ["read", "create", "update", "delete"],
   submissions: ["read", "read_all", "create", "update", "delete"],
   settings: ["read", "update"],
+  deleted_entries: ["read"],
+  uke_permits: ["read"],
+  uke_permits_orphaned: ["read"],
 });