+++ packages/drizzle/db.ts
@@ -12,7 +12,7 @@ if (!connectionString) {
 export const sql = postgres(connectionString, {
   prepare: false,
   max: 10,
-  idle_timeout: 20,
+  idle_timeout: 100,
   connect_timeout: 10,
 });